requirements.txt
accelerate==0.24.1
av==11.0.0
basicsr==1.4.2
black==24.4.2
caffe2==0.8.1
cityscapesscripts==2.2.3
cloudpickle==3.0.0
diffusers==0.27.2
docutils==0.21.2
einops==0.8.0
external==0.0.1
fairscale==0.4.13
fvcore==0.1.5.post20221221
huggingface_hub==0.20.2
hydra-core==1.3.2
imageio==2.33.1
iopath==0.1.9
matplotlib==3.8.2
mediapipe==0.10.9
mmcv==2.1.0
mmdet==3.3.0
mmdet==3.2.0
mmpose==1.3.1
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.18.0
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
openvino==2024.1.0
Pillow==9.5.0
Pillow==10.3.0
psutil==5.9.7
pycocotools==2.0.7
pyfacer==0.0.4
Pygments==2.18.0
pytest==8.2.1
pytorch_lightning==2.2.4
PyYAML==6.0.1
PyYAML==6.0.1
recommonmark==0.7.1
refile==8.0.2
Requests==2.32.2
safetensors==0.4.3
scipy==1.13.0
setuptools==60.2.0
Shapely==2.0.4
skimage==0.0
Sphinx==7.3.7
sphinx_rtd_theme==2.0.0
tabulate==0.9.0
termcolor==2.4.0
timm==0.9.12
torch==2.0.0
torchvision==0.15.1+cu117
tqdm==4.65.2
transformers==4.31.0
xformers==0.0.19
